+++ androidauto/build.gradle
@@ -30,12 +30,6 @@ android {
         consumerProguardFiles 'proguard-rules.pro', "../proguard/proguard-project.pro"
     }
 
-    buildTypes {
-        release {
-            crunchPngs = false
-        }
-    }
-
     testOptions {
         execution 'ANDROIDX_TEST_ORCHESTRATOR'
     }
