--- libnavigation-base/src/main/java/com/mapbox/navigation/base/internal/route/RouteRefreshMetadataFactory.kt
@@ -1,11 +1,0 @@
-package com.mapbox.navigation.base.internal.route
-
-import com.mapbox.navigation.base.ExperimentalMapboxNavigationAPI
-import com.mapbox.navigation.base.route.RouteRefreshMetadata
-
-@ExperimentalMapboxNavigationAPI
-fun createRouteRefreshMetadata(
-    isUpToDate: Boolean,
-) = RouteRefreshMetadata(
-    isUpToDate = isUpToDate,
-)
