+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/route/NavigationRoute.kt
@@ -57,8 +57,7 @@ import kotlin.time.Duration.Companion.milliseconds
  * Prefer using [waypoints] instead of [DirectionsRoute.waypoints] from [directionsRoute].
  * @param responseOriginAPI describes API which generated data for [NavigationRoute].
  * @param overriddenTraffic describes a segment of [NavigationRoute] with overridden
- * @param routeRefreshMetadata contains data that describes refreshing of this route object.
- * The field is null until the first refresh.
+ * congestion_numeric.
  */
 class NavigationRoute private constructor(
     val directionsRoute: DirectionsRoute,
@@ -71,8 +70,6 @@ class NavigationRoute private constructor(
     @ExperimentalMapboxNavigationAPI
     @ResponseOriginAPI
     val responseOriginAPI: String,
-    @ExperimentalMapboxNavigationAPI
-    val routeRefreshMetadata: RouteRefreshMetadata?,
 ) {
 
     internal constructor(
@@ -95,7 +92,6 @@ class NavigationRoute private constructor(
         expirationTimeElapsedSeconds,
         overriddenTraffic,
         responseOriginAPI,
-        null,
     )
 
     /**
@@ -156,7 +152,6 @@ class NavigationRoute private constructor(
             responseOriginAPI,
             responseUUID,
             expirationTimeElapsedSeconds,
-            routeRefreshMetadata,
         )
         return gson.toJson(state)
     }
@@ -192,7 +187,6 @@ class NavigationRoute private constructor(
                     state.expirationTimeElapsedSeconds,
                     responseOriginAPI = state.responseOriginAPI,
                     overriddenTraffic = null,
-                    routeRefreshMetadata = state.routeRefreshMetadata,
                 )
                 ExpectedFactory.createValue(route)
             } catch (t: Throwable) {
@@ -631,7 +625,6 @@ class NavigationRoute private constructor(
         nativeRoute: RouteInterface = this.nativeRoute,
         overriddenTraffic: CongestionNumericOverride? = this.overriddenTraffic,
         expirationTimeElapsedSeconds: Long? = this.expirationTimeElapsedSeconds,
-        routeRefreshMetadata: RouteRefreshMetadata? = this.routeRefreshMetadata,
     ): NavigationRoute = NavigationRoute(
         directionsRoute,
         waypoints,
@@ -641,7 +634,6 @@ class NavigationRoute private constructor(
         expirationTimeElapsedSeconds,
         overriddenTraffic,
         this.responseOriginAPI,
-        routeRefreshMetadata,
     )
 
     @Keep
@@ -656,7 +648,6 @@ class NavigationRoute private constructor(
         val responseOriginAPI: String,
         val responseUUID: String,
         val expirationTimeElapsedSeconds: Long?,
-        val routeRefreshMetadata: RouteRefreshMetadata?,
     )
 }
