+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/routerefresh/ExpiringDataRemover.kt
@@ -2,8 +2,6 @@ package com.mapbox.navigation.core.routerefresh
 
 import com.mapbox.api.directions.v5.models.DirectionsRoute
 import com.mapbox.api.directions.v5.models.RouteLeg
-import com.mapbox.navigation.base.ExperimentalMapboxNavigationAPI
-import com.mapbox.navigation.base.internal.route.createRouteRefreshMetadata
 import com.mapbox.navigation.base.internal.route.update
 import com.mapbox.navigation.base.internal.time.parseISO8601DateToLocalTimeOrNull
 import com.mapbox.navigation.base.route.NavigationRoute
@@ -37,7 +35,6 @@ internal class ExpiringDataRemover(
         )
     }
 
-    @OptIn(ExperimentalMapboxNavigationAPI::class)
     private fun removeExpiringDataFromRoute(
         route: NavigationRoute,
         currentLegIndex: Int,
@@ -58,7 +55,6 @@ internal class ExpiringDataRemover(
         return route.update(
             directionsRouteBlock = directionsRouteBlock,
             waypointsBlock = { this },
-            routeRefreshMetadata = createRouteRefreshMetadata(isUpToDate = false),
         )
     }
