+++ libnavigation-core/src/test/java/com/mapbox/navigation/core/internal/congestions/processor/DecreaseTrafficUpdateActionHandlerTest.kt
@@ -1,5 +1,3 @@
-@file:OptIn(ExperimentalMapboxNavigationAPI::class)
-
 package com.mapbox.navigation.core.internal.congestions.processor
 
 import com.mapbox.api.directions.v5.models.LegAnnotation
@@ -8,7 +6,6 @@ import com.mapbox.api.directions.v5.models.ManeuverModifier
 import com.mapbox.api.directions.v5.models.RouteLeg
 import com.mapbox.api.directions.v5.models.StepIntersection
 import com.mapbox.api.directions.v5.models.StepManeuver
-import com.mapbox.navigation.base.ExperimentalMapboxNavigationAPI
 import com.mapbox.navigation.base.internal.CongestionNumericOverride
 import com.mapbox.navigation.base.internal.route.overriddenTraffic
 import com.mapbox.navigation.base.internal.route.update
