+++ libnavui-maps/src/main/java/com/mapbox/navigation/ui/maps/route/callout/api/MapboxRouteCalloutApi.kt
@@ -136,6 +136,8 @@ class MapboxRouteCalloutApi(
         val primaryRoute = routes.firstOrNull() ?: return emptyList()
         val alternativeRoutes = routes.drop(1)
 
+        if (alternativeRoutes.isEmpty()) return emptyList()
+
         return when (options.routeCalloutType) {
             RouteCalloutType.RouteDurations -> createRoutePreviewCallouts(
                 primaryRoute,
@@ -202,7 +204,7 @@ class MapboxRouteCalloutApi(
         val durationDiffAbsoluteValue = durationDiff.absoluteValue
 
         return durationDiffAbsoluteValue to when {
-            durationDiffAbsoluteValue <= options.similarDurationDelta -> DurationDifferenceType.Same
+            durationDiffAbsoluteValue < options.similarDurationDelta -> DurationDifferenceType.Same
 
             durationDiff < 0.seconds -> DurationDifferenceType.Slower
 
@@ -215,17 +217,24 @@ class MapboxRouteCalloutApi(
         primaryRoutePoints: List<Point>,
         alternativeRoutes: List<NavigationRoute>,
     ): List<RouteCallout.Eta> {
-        return buildList(capacity = alternativeRoutes.size + 1) {
-            val primaryGeometry = LineString.fromLngLats(primaryRoutePoints)
-            add(RouteCallout.Eta(primaryRoute, primaryGeometry, isPrimary = true))
+        val primaryCallout = RouteCallout.Eta(
+            primaryRoute,
+            LineString.fromLngLats(primaryRoutePoints),
+            isPrimary = true,
+        )
 
-            alternativeRoutes.mapTo(destination = this) { alternativeRoute ->
-                val geometry = alternativesGeometryDifference[alternativeRoute.id]
-                    ?: alternativeRoute.directionsRoute.completeGeometryToLineString()
+        val alternativeCallouts = alternativeRoutes.map {
+            val geometry = alternativesGeometryDifference[it.id]
+                ?: it.directionsRoute.completeGeometryToLineString()
 
-                RouteCallout.Eta(alternativeRoute, geometry, isPrimary = false)
-            }
+            RouteCallout.Eta(
+                it,
+                geometry,
+                isPrimary = false,
+            )
         }
+
+        return listOf(primaryCallout) + alternativeCallouts
     }
 
     private companion object {
